+++ pallets/programs/src/lib.rs
@@ -96,6 +96,8 @@ pub mod pallet {
     pub struct ProgramInfo<AccountId> {
         /// The bytecode of the program.
         pub bytecode: Vec<u8>,
+        /// The type definition of the program
+        pub program_type_definition: Vec<u8>,
         /// Owners of the program
         pub program_modification_account: AccountId,
         /// Accounts that use this program
@@ -129,6 +131,9 @@ pub mod pallet {
 
             /// The new program hash.
             program_hash: T::Hash,
+
+            /// The new program type definition
+            program_type_definition: Vec<u8>,
         },
         /// The bytecode of a program was removed.
         ProgramRemoved {
@@ -163,10 +168,17 @@ pub mod pallet {
         /// Note that the caller becomes the program-modification account.
         #[pallet::call_index(0)]
         #[pallet::weight({<T as Config>::WeightInfo::set_program()})]
-        pub fn set_program(origin: OriginFor<T>, new_program: Vec<u8>) -> DispatchResult {
+        pub fn set_program(
+            origin: OriginFor<T>,
+            new_program: Vec<u8>,
+            program_type_definition: Vec<u8>,
+        ) -> DispatchResult {
             let program_modification_account = ensure_signed(origin)?;
-            let program_hash = T::Hashing::hash(&new_program);
-            let new_program_length = new_program.len();
+            let mut hash_input = vec![];
+            hash_input.extend(&new_program);
+            hash_input.extend(&program_type_definition);
+            let program_hash = T::Hashing::hash(&hash_input);
+            let new_program_length = new_program.len() + program_type_definition.len();
             ensure!(
                 new_program_length as u32 <= T::MaxBytecodeLength::get(),
                 Error::<T>::ProgramLengthExceeded
@@ -179,6 +191,7 @@ pub mod pallet {
                 program_hash,
                 &ProgramInfo {
                     bytecode: new_program.clone(),
+                    program_type_definition: program_type_definition.clone(),
                     program_modification_account: program_modification_account.clone(),
                     ref_counter: 0u128,
                 },
@@ -195,6 +208,7 @@ pub mod pallet {
             Self::deposit_event(Event::ProgramCreated {
                 program_modification_account,
                 program_hash,
+                program_type_definition,
             });
             Ok(())
         }
@@ -218,7 +232,7 @@ pub mod pallet {
             ensure!(old_program_info.ref_counter == 0, Error::<T>::ProgramInUse);
             Self::unreserve_program_deposit(
                 &old_program_info.program_modification_account,
-                old_program_info.bytecode.len(),
+                old_program_info.bytecode.len() + old_program_info.program_type_definition.len(),
             );
             let mut owned_programs_length = 0;
             OwnedPrograms::<T>::try_mutate(